Optimise Vertex Pos displacement shader
Instead of using a for loop to find the closest Hex center we can
calculate what the closest one should be, by using the row and
column offsets and the relative positioning of our vertices.

We can also remove the need for the shader to know the grid dimensions,
i.e. we do not need to modulo by the number of hexagons in each row
because we always calculate how many 'hexagons' the vertex is from the origin.

We can dynamically grow the honeycomb and the vertices should always be able
to figure out what hexagon they belong to.